@@ -3,47 +3,37 @@ import locales from '../locales'
 
 export const ThemeContext = React.createContext('hiui-blue')
 export const LocaleContext = React.createContext('zh-CN')
-export const VersionContext = React.createContext(2)
 
 export default (WrappedComponent) => {
   class WrapperComponent extends Component {
     static displayName = WrappedComponent.name
     render () {
-      const { theme, locale, version, innerRef, ...restProps } = this.props
+      const { theme, locale, innerRef, ...restProps } = this.props
       let ConsumerComponent = (
         <ThemeContext.Consumer>
           {(contextTheme) => (
             <LocaleContext.Consumer>
               {(contextLocale) => (
-                <VersionContext.Consumer>
-                  {(contextVersion) => (
-                    <WrappedComponent
-                      theme={contextTheme}
-                      version={contextVersion}
-                      locale={contextLocale}
-                      localeDatas={locales[contextLocale]}
-                      ref={innerRef}
-                      innerRef={innerRef}
-                      {...restProps}
-                    />
-                  )}
-                </VersionContext.Consumer>
+                <WrappedComponent
+                  theme={contextTheme}
+                  locale={contextLocale}
+                  localeDatas={locales[contextLocale]}
+                  ref={innerRef}
+                  innerRef={innerRef}
+                  {...restProps}
+                />
               )}
             </LocaleContext.Consumer>
           )}
         </ThemeContext.Consumer>
       )
-
       return wrapProvider(theme, ThemeContext)(locale, LocaleContext)(
-        version,
-        VersionContext
-      )(ConsumerComponent)
+        ConsumerComponent
+      )
     }
   }
   return forwardRef((props, ref) => {
-    return (
-      <WrapperComponent {...props} innerRef={ref} />
-    )
+    return <WrapperComponent {...props} innerRef={ref} />
   })
 }
 
