@@ -6,11 +6,8 @@ import {formatterDate, FORMATS} from './constants'
 import PropTypes from 'prop-types'
 import DatePickerType from './Type'
 
-import {parse, dateFormat, isValid, startOfWeek, endOfWeek} from './dateUtil'
+import { dateFormat, isValid, startOfWeek, endOfWeek, toDate, parse } from './dateUtil'
 class BasePicker extends Component {
-  inputRoot = null
-  input = null
-  rInput = null
   constructor (props) {
     super(props)
     this.state = {
@@ -24,6 +21,9 @@ class BasePicker extends Component {
       rightPlaceholder: '',
       format: ''
     }
+    this.inputRoot = React.createRef()
+    this.input = null
+    this.rInput = null
   }
   setPlaceholder () {
     const {placeholder, localeDatas, type} = this.props
@@ -43,52 +43,10 @@ class BasePicker extends Component {
       rightPlaceholder
     })
   }
-  _parseProps (props, callback) {
-    let {value, defaultValue, showTime, type, format, localeDatas, weekOffset, timeInterval = 240} = props
-    format = format || FORMATS[type]
-    let _value = value || defaultValue
-    let start
-    let end
-    let date
-    let leftText = ''
-    let rightText = ''
-    if (_value) {
-      if (Object.prototype.toString.call(_value) === '[object Object]') {
-        start = _value.start || null
-        end = _value.end || new Date()
-      } else {
-        start = _value
-        if (type.includes('range')) {
-          end = parse(start)
-          if (type === 'weekrange') {
-            start = startOfWeek(start)
-            end = endOfWeek(end)
-          }
-        }
-      }
-
-      if (type === 'timeperiod' && isValid(start)) {
-        let startTime = start.getTime()
-        startTime += timeInterval * 60 * 1000
-        end = new Date(startTime)
-      }
-    }
-    date = {
-      startDate: parse(start),
-      endDate: parse(end)
-    }
-    leftText = isValid(date.startDate) ? formatterDate(type, date.startDate, format, showTime, localeDatas, weekOffset) : ''
-    rightText = isValid(date.endDate) ? formatterDate(type, date.endDate, format, showTime, localeDatas, weekOffset) : ''
-    this.setState({
-      texts: [leftText, rightText],
-      date,
-      format
-    })
-  }
   componentDidMount () {
     this._parseProps(this.props)
     this.setPlaceholder()
-    let rect = this.inputRoot.getBoundingClientRect()
+    let rect = this.inputRoot.current.getBoundingClientRect()
     this.calcPanelPos(rect)
   }
   calcPanelPos (rect) {
@@ -122,6 +80,53 @@ class BasePicker extends Component {
       this._parseProps(nextProps)
     }
   }
+  compatibleFormatString (format) {
+    return format.replace(/[Y+|D+]/g, (word) => {
+      return word.toLowerCase()
+    })
+  }
+  _parseProps (props) {
+    let {value, defaultValue, showTime, type, format, localeDatas, weekOffset, timeInterval = 240} = props
+    format = this.compatibleFormatString(format || FORMATS[type])
+    let _value = value || defaultValue
+    let start
+    let end
+    let date
+    let leftText = ''
+    let rightText = ''
+    if (_value) {
+      if (Object.prototype.toString.call(_value) === '[object Object]') {
+        start = toDate(_value.start) || null
+        end = toDate(_value.end) || new Date()
+      } else {
+        start = toDate(_value)
+        if (type.includes('range')) {
+          end = toDate(start)
+          if (type === 'weekrange') {
+            start = startOfWeek(start)
+            end = endOfWeek(end)
+          }
+        }
+      }
+
+      if (type === 'timeperiod' && isValid(start)) {
+        let startTime = start.getTime()
+        startTime += timeInterval * 60 * 1000
+        end = new Date(startTime)
+      }
+    }
+    date = {
+      startDate: toDate(start),
+      endDate: toDate(end)
+    }
+    leftText = isValid(date.startDate) ? formatterDate(type, date.startDate, format, showTime, localeDatas, weekOffset) : ''
+    rightText = isValid(date.endDate) ? formatterDate(type, date.endDate, format, showTime, localeDatas, weekOffset) : ''
+    this.setState({
+      texts: [leftText, rightText],
+      date,
+      format
+    })
+  }
   onPick (date, showPanel) {
     if (!date.startDate) {
       date = {startDate: date, endDate: undefined}
@@ -131,8 +136,7 @@ class BasePicker extends Component {
     this.setState({
       date,
       texts: [formatterDate(type, date.startDate, format, showTime, localeDatas, weekOffset), formatterDate(type, date.endDate, format, showTime, localeDatas, weekOffset)],
-      showPanel,
-      isFocus: false
+      showPanel
     }, () => {
       if (!showPanel) {
         this.callback()
@@ -151,6 +155,10 @@ class BasePicker extends Component {
         return
       }
 
+      if (type === 'time') {
+        onChange(startDate, dateFormat(startDate, this.state.format))
+        return
+      }
       if (['timerange', 'timeperiod', 'daterange'].includes(type)) {
         onChange({start: startDate, end: endDate})
         return
@@ -177,27 +185,18 @@ class BasePicker extends Component {
     }
   }
   timeCancel () {
-    const {tempDate, format} = this.state
+    const {format, date} = this.state
     const {type, showTime, localeDatas, weekOffset} = this.props
-    if (tempDate) {
-      this.setState({
-        date: new Date(tempDate),
-        text: formatterDate(type, new Date(tempDate), format, showTime, localeDatas, weekOffset),
-        showPanel: false
-      })
-    } else {
-      this.setState({
-        showPanel: false,
-        isFocus: false
-      })
-    }
+    this.setState({
+      showPanel: false,
+      texts: [formatterDate(type, date.startDate || date, format, showTime, localeDatas, weekOffset), formatterDate(type, date.endDate, format, showTime, localeDatas, weekOffset)],
+      isFocus: false
+    })
   }
   inputChangeEvent () {
-    let { texts, date } = this.state
-    // const {type, showTime, localeDatas, weekOffset} = this.props
-    let startDate = parse(texts[0])
-    let endDate = parse(texts[1])
-
+    let { texts, date, format } = this.state
+    let startDate = parse(texts[0], format, new Date())
+    let endDate = parse(texts[1], format, new Date())
     if (startDate && isValid(startDate)) {
       date.startDate ? date.startDate = startDate : date = startDate
     }
@@ -212,15 +211,12 @@ class BasePicker extends Component {
     if (tar.className.indexOf('clear') !== -1) {
       this.setState({
         texts: ['', ''],
-        showPanel: false,
-        date: null
+        showPanel: false
       })
       return false
     }
     if (tar !== this.input && tar !== this.rInput) {
       this.timeCancel()
-    }
-    if (!this.showPanel && tar !== this.input && tar !== this.rInput) {
       this.callback()
     }
   }
@@ -247,7 +243,7 @@ class BasePicker extends Component {
             showPanel: true,
             isFocus: true
           })
-          this.calcPanelPos(this.inputRoot.getBoundingClientRect())
+          this.calcPanelPos(this.inputRoot.current.getBoundingClientRect())
         }}
         value={text}
       />
@@ -257,7 +253,7 @@ class BasePicker extends Component {
     const {onChange, type} = this.props
     if (onChange) {
       onChange(
-        type.includes('range') || type === 'timeperiod' ? {start: '', end: ''} : ''
+        (type.includes('range') || type === 'timeperiod') ? {start: '', end: ''} : ''
       )
     }
     this.setState({date: {startDate: null, endDate: null}, texts: ['', ''], isFocus: false})
@@ -268,13 +264,13 @@ class BasePicker extends Component {
     const iconCls = classNames(
       'hi-datepicker__input-icon',
       'hi-icon',
-      (isFocus && clearable) ? 'icon-close-circle clear' : ((showTime || type === 'timeperiod') ? 'icon-time' : 'icon-date')
+      (isFocus && clearable) ? 'icon-close-circle clear' : ((type.includes('time') || showTime) ? 'icon-time' : 'icon-date')
     )
     return (isFocus && clearable)
       ? <span className={iconCls} onClick={this._clear.bind(this)} />
       : <span className={iconCls} onClick={(e) => {
         if (this.props.disabled) return
-        this.calcPanelPos(this.inputRoot.getBoundingClientRect())
+        this.calcPanelPos(this.inputRoot.current.getBoundingClientRect())
         this.setState({showPanel: true, isFocus: true})
       }} />
   }
@@ -294,7 +290,7 @@ class BasePicker extends Component {
     return (
       <div className={_cls}>
         {this._input(this.state.texts[0], 'input', this.state.leftPlaceholder)}
-        <span>{localeDatas.datePicker.to}</span>
+        <span className='hi-datepicker__input--connection'>{localeDatas.datePicker.to}</span>
         {this._input(this.state.texts[1], 'rInput', this.state.rightPlaceholder)}
         {this._icon()}
       </div>
@@ -321,7 +317,7 @@ class BasePicker extends Component {
   render () {
     const {type, showTime} = this.props
     return (
-      <span ref={el => { this.inputRoot = el }} className='hi-datepicker__input-root'>
+      <span ref={this.inputRoot} className='hi-datepicker__input-root'>
         {
           (type.indexOf('range') !== -1 || type === 'timeperiod') ? this.renderRangeInput() : this.renderNormalInput()
         }
@@ -347,13 +343,13 @@ BasePicker.propTypes = {
       return null
     }
     if (val.start && val.end) {
-      const _start = dateFormat(val.start)
-      const _end = dateFormat(val.end)
-      if (_start === 'Invalid Date' || _end === 'Invalid Date') {
+      const _start = isValid(val.start)
+      const _end = isValid(val.end)
+      if (!_start || !_end) {
         return new Error(`Invalid prop ${propName} supplied to ${componentName}. Validation failed. start or end is an invalid date.`)
       }
     } else {
-      if (dateFormat(val) === 'Invalid Date') {
+      if (!isValid(val)) {
         return new Error(`Invalid prop ${propName} supplied to ${componentName}. Validation failed. value is an invalid data.`)
       }
     }
