@@ -4,6 +4,8 @@ import BasePicker from './BasePicker'
 import DatePanel from './DatePanel'
 import DateRangePanel from './DateRangePanel'
 import WeekRangePanel from './WeekRangePanel'
+import Provider from '../context'
+
 class DatePicker extends BasePicker {
   initPanel (state, props) {
     let component = null
@@ -62,4 +64,4 @@ class DatePicker extends BasePicker {
     return component
   }
 }
-export default DatePicker
+export default Provider(DatePicker)
