@@ -1,6 +1,5 @@
 import React, {Component} from 'react'
 import classNames from 'classnames'
-import Provider from '../context'
 import {getHours, getMinutes} from './dateUtil'
 class TimePeriodPanel extends Component {
   constructor (props) {
@@ -63,4 +62,4 @@ class TimePeriodPanel extends Component {
   }
 }
 
-export default Provider(TimePeriodPanel)
+export default TimePeriodPanel
