@@ -17,13 +17,13 @@ class TimeRangePanel extends Component {
     this.props.onPick(r, bol)
   }
   render () {
-    const { localeDatas, date, style } = this.props
+    const { date, style } = this.props
     const {startDate, endDate} = date
     return (
       <div className='hi-timepicker hi-timepicker--timerange' style={style}>
-        <Time localeDatas={localeDatas} date={startDate} onPick={this.onTimePick.bind(this, 'left')} onlyTime />
+        <Time {...this.props} date={startDate} onPick={this.onTimePick.bind(this, 'left')} onlyTime />
         <div className='hi-timepicker__split' />
-        <Time localeDatas={localeDatas} date={endDate} onPick={this.onTimePick.bind(this, 'right')} onlyTime disableTime={startDate} />
+        <Time {...this.props} date={endDate} onPick={this.onTimePick.bind(this, 'right')} onlyTime disableTime={startDate} />
       </div>
     )
   }
