@@ -1,36 +1,44 @@
-import getDaysInMonth from 'date-fns/get_days_in_month'
-import subMonths from 'date-fns/sub_months'
-import getDay from 'date-fns/get_day'
-import startOfMonth from 'date-fns/start_of_month'
-import isWithinRange from 'date-fns/is_within_range'
-import isSameDay from 'date-fns/is_same_day'
-import compareAsc from 'date-fns/compare_asc'
-import startOfDay from 'date-fns/start_of_day'
-import endOfDay from 'date-fns/end_of_day'
+import getDaysInMonth from 'date-fns/getDaysInMonth'
+import subMonths from 'date-fns/subMonths'
+import getDay from 'date-fns/getDay'
+import startOfMonth from 'date-fns/startOfMonth'
+import isWithinInterval from 'date-fns/isWithinInterval'
+import isSameDay from 'date-fns/isSameDay'
+import compareAsc from 'date-fns/compareAsc'
+import startOfDay from 'date-fns/startOfDay'
+import endOfDay from 'date-fns/endOfDay'
 import parse from 'date-fns/parse'
-import startOfWeek from 'date-fns/start_of_week'
-import endOfWeek from 'date-fns/end_of_week'
+import toDate from 'date-fns/toDate'
+import startOfWeek from 'date-fns/startOfWeek'
+import endOfWeek from 'date-fns/endOfWeek'
 import dateFormat from 'date-fns/format'
-import addMonths from 'date-fns/add_months'
-import isSameMonth from 'date-fns/is_same_month'
-import getYear from 'date-fns/get_year'
-import getMonth from 'date-fns/get_month'
-import isToday from 'date-fns/is_today'
-import getHours from 'date-fns/get_hours'
-import getMinutes from 'date-fns/get_minutes'
-import getSeconds from 'date-fns/get_seconds'
-import addHours from 'date-fns/add_hours'
-import subDays from 'date-fns/sub_days'
-import differenceInDays from 'date-fns/difference_in_days'
-import dfIsValid from 'date-fns/is_valid'
-import addYears from 'date-fns/add_years'
-import subYears from 'date-fns/sub_years'
+import addMonths from 'date-fns/addMonths'
+import isSameMonth from 'date-fns/isSameMonth'
+import getYear from 'date-fns/getYear'
+import getMonth from 'date-fns/getMonth'
+import isToday from 'date-fns/isToday'
+import getHours from 'date-fns/getHours'
+import getMinutes from 'date-fns/getMinutes'
+import getSeconds from 'date-fns/getSeconds'
+import addHours from 'date-fns/addHours'
+import subDays from 'date-fns/subDays'
+import differenceInDays from 'date-fns/differenceInDays'
+import dfIsValid from 'date-fns/isValid'
+import addYears from 'date-fns/addYears'
+import subYears from 'date-fns/subYears'
+import parseISO from 'date-fns/parseISO'
 
 const isValid = (date) => {
-  return (date && dfIsValid(parse(date)))
+  if (typeof date === 'string') {
+    return dfIsValid(parseISO(date))
+  }
+  if (typeof date === 'number') {
+    return dfIsValid(toDate(date))
+  }
+  return (date && dfIsValid(toDate(date)))
 }
 const getValidDate = (date) => {
-  return isValid(date) ? parse(date) : new Date()
+  return isValid(date) ? toDate(date) : new Date()
 }
 const getStartDate = (dateObj) => {
   return getValidDate(dateObj.startDate)
@@ -47,7 +55,7 @@ export {
   subYears, // 年份减
   getDay, // 获取周几
   startOfMonth, // 指定日期月份的1日
-  isWithinRange, // 是否在指定日期范围内
+  isWithinInterval, // 是否在指定日期范围内
   isSameDay, // 是否是同一天（忽略时分秒）
   compareAsc, // 比较两个日期
   startOfDay, // 一天的开始
@@ -69,5 +77,7 @@ export {
   isValid, // 是否有效时间
   getStartDate, // 封装 - 获取开始时间
   getEndDate, // 封装 - 获取结果时间
-  getValidDate // 获取有效的时间
+  getValidDate, // 获取有效的时间
+  toDate,
+  parseISO
 }
