@@ -6,7 +6,7 @@ import Icon from '../../icon'
 import classNames from 'classnames'
 import Provider from '../../context'
 import TimePeriodPanel from './TimePeriodPanel'
-import {dateFormat, parse, addHours} from './dateUtil'
+import {dateFormat, parseISO, addHours} from './dateUtil'
 
 class DatePanel extends Component {
   constructor (props) {
@@ -193,9 +193,9 @@ class DatePanel extends Component {
     }
   }
   onTimePeriodPick (periodS, periodE) {
-    const currentDate = dateFormat(this.state.date, 'YYYY-MM-DD')
-    parse(`${currentDate} ${periodS}`)
-    this.props.onPick({startDate: parse(`${currentDate} ${periodS}`), endDate: parse(`${currentDate} ${periodE}`)}, true)
+    const currentDate = dateFormat(this.state.date, 'yyyy-MM-dd')
+    parseISO(`${currentDate} ${periodS}`)
+    this.props.onPick({startDate: parseISO(`${currentDate} ${periodS}`), endDate: parseISO(`${currentDate} ${periodE}`)}, true)
   }
   onTimePick (date, bol) {
     this.setState({
