@@ -1,4 +1,7 @@
+import zhCN from './zh-CN'
+import enUS from './en-US'
+
 export default {
-  'zh-CN': require('./zh-CN'),
-  'en-US': require('./en-US')
+  'zh-CN': zhCN,
+  'en-US': enUS
 }
