@@ -5,6 +5,12 @@ const prefix = 'counter-base'
 const code = `import React from 'react'
 import Counter from '@hi-ui/hiui/es/counter'\n
 class Demo extends React.Component {
+  constructor(props){
+    super(props)
+    this.state={
+      value:4
+    }
+  }
   render() {
     return (
       <div>
@@ -16,11 +22,11 @@ class Demo extends React.Component {
         />
         <p />
         <Counter
-          defaultValue='8888'
+          value={this.state.value}
           step='2'
           min='2'
           max='8'
-          onChange={(e, val) => {console.log('----val', e, val)}}
+          onChange={(e, val) => {this.setState({value:val})}}
         />
       </div>
     )
