@@ -1,9 +1,9 @@
 import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
-import Dropdown from '../../../components/dropdown'
-const prefix = 'dropdown-trigger'
-const rightOptions = ['鼠标悬停', '左键单击', '右键单击', '多种组合']
-const desc = '通过 trigger 属性切换菜单触发方式'
+import { Dropdown, Message } from '../../../components'
+const prefix = 'dropdown-type'
+const rightOptions = ['文字型', '按钮型', '按钮组']
+const desc = '通过 type 属性切换按钮类型'
 
 const code = [
   {
@@ -34,14 +34,48 @@ class Demo extends React.Component {
     return (
       <Dropdown
         data={list}
-        trigger='hover'
+        title="鼠标悬停"
+      />
+    )
+  }
+}`,
+    opt: ['文字型']
+  },
+  {
+    code: `import React from 'react'
+import Icon from '@hi-ui/hiui/es/icon'
+import Dropdown from '@hi-ui/hiui/es/dropdown'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      list: [{
+        title: '菜单一',
+      }, {
+        title: '菜单二'
+      }, {
+        title: '菜单三'
+      }, {
+        title: '-'
+      }, {
+        title: '链接一',
+        href: 'https://www.mi.com',
+        disabled: true
+      }]
+    }
+  }
+  render () {
+    const { list } = this.state
+    return (
+      <Dropdown
+        data={list}
+        type='button'
         title='鼠标悬停'
-        type='button'
       />
     )
   }
 }`,
-    opt: ['鼠标悬停']
+    opt: ['按钮型']
   },
   {
     code: `import React from 'react'
@@ -71,95 +105,23 @@ class Demo extends React.Component {
     return (
       <Dropdown
         data={list}
-        type='button'
-        trigger='click'
-        title='左键单击'
+        type='group'
+        title='点击'
+        onButtonClick={() => {
+          Message.open({ title: '点击左侧按钮' })
+        }}
       />
     )
   }
 }`,
-    opt: ['左键单击']
-  },
-  {
-    code: `import React from 'react'
-import Icon from '@hi-ui/hiui/es/icon'
-import Dropdown from '@hi-ui/hiui/es/dropdown'\n
-class Demo extends React.Component {
-  constructor () {
-    super()
-    this.state = {
-      list: [{
-        title: '菜单一',
-      }, {
-        title: '菜单二'
-      }, {
-        title: '菜单三'
-      }, {
-        title: '-'
-      }, {
-        title: '链接一',
-        href: 'https://www.mi.com',
-        disabled: true
-      }]
-    }
-  }
-  render () {
-    const { list } = this.state
-    return (
-      <Dropdown
-        data={list}
-        type='button'
-        trigger='contextmenu'
-        title='右键单击'
-      />
-    )
-  }
-}`,
-    opt: ['右键单击']
-  },
-  {
-    code: `import React from 'react'
-import Icon from '@hi-ui/hiui/es/icon'
-import Dropdown from '@hi-ui/hiui/es/dropdown'\n
-class Demo extends React.Component {
-  constructor () {
-    super()
-    this.state = {
-      list: [{
-        title: '菜单一',
-      }, {
-        title: '菜单二'
-      }, {
-        title: '菜单三'
-      }, {
-        title: '-'
-      }, {
-        title: '链接一',
-        href: 'https://www.mi.com',
-        disabled: true
-      }]
-    }
-  }
-  render () {
-    const { list } = this.state
-    return (
-      <Dropdown
-        data={list}
-        trigger={['click', 'contextmenu']}
-        type='button'
-        title='左键或右键单击'
-      />
-    )
-  }
-}`,
-    opt: ['多种组合']
+    opt: ['按钮组']
   }
 ]
 
 const DemoBase = () => (
   <DocViewer
     code={code}
-    scope={{ Dropdown }}
+    scope={{ Dropdown, Message }}
     prefix={prefix}
     rightOptions={rightOptions}
     desc={desc}
